Amoxtli/requirements.txt
astroid==2.0.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
dj-database-url==0.5.0
django>=2.1.11
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-mysql==3.0.0.post1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.11.0
urllib3>=1.24.2
whitenoise==4.1.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.1